requirements.txt
streamlit>=1.35.0
Pillow>=10.0.0
pandas>=2.0.0
